# docs/swagger.yaml
definitions:
...
  models.News:
    properties:
      date:
        type: string
      id:
        type: integer
      image_url:
        type: string
      isi_konten:
        type: string
      kategori:
        enum:
        - News
        - Event
        type: string
      link:
        type: string
      nama_penulis:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    required:
    - kategori
    - nama_penulis
    - title
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /news:
    get:
      description: Retrieves all news from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.News'
            type: array
      summary: Get All News
      tags:
      - News
    post:
      consumes:
      - multipart/form-data
      description: Insert a news and saves them to the database.
      parameters:
      - description: Judul news
        in: formData
        name: title
        required: true
        type: string
      - description: Kategori news
        in: formData
        name: kategori
        required: true
        type: string
      - description: Thumbnail news
        in: formData
        name: thumbnail
        required: true
        type: file
      - description: Isi konten news
        in: formData
        name: isi_konten
        required: true
        type: string
      - description: Nama penulis news
        in: formData
        name: nama_penulis
        required: true
        type: string
      - description: Link news
        in: formData
        name: link
        required: true
        type: string
      - description: Date news
        in: formData
        name: date
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: Insert a new news
      tags:
      - News
  /news/{id}:
    delete:
      description: Delete a News and its thumbnail from the database and storage.
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: News deleted successfully
          schema:
            type: string
      summary: Delete News
      tags:
      - News
    get:
      description: Retrieves news data by its ID.
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get News By ID
      tags:
      - News
    put:
      consumes:
      - multipart/form-data
      description: Edits a News by its ID
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: string
      - description: Title News
        in: formData
        name: title
        required: true
        type: string
      - description: Kategori News
        in: formData
        name: kategori
        required: true
        type: string
      - description: Thumbnail News
        in: formData
        name: thumbnail
        type: file
      - description: Isi Konten News
        in: formData
        name: isi_konten
        required: true
        type: string
      - description: Nama Penulis News
        in: formData
        name: nama_penulis
        required: true
        type: string
      - description: Link News
        in: formData
        name: link
        required: true
        type: string
      - description: Date News
        in: formData
        name: date
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: Edit News
      tags:
      - News
  /news/category/{category}:
    get:
      description: Retrieves news data by its Category.
      parameters:
      - description: News Category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get All News By Category
      tags:
      - News
  /news/file/{id}:
    get:
      description: Retrieves the image of a News by its ID.
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get Thumbnail News
      tags:
      - News
...
swagger: "2.0"